src/main/bin/requirements.txt
requests
PyMySQL
commandr